pyproject.toml
[tool.poetry]
name = "yl_user_auth"
version = "0.1.0"
description = ""
authors = ["Alexander Lyotov <sippie84@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
binary = "^1.0.0"
matches = "^0.0.1"
anyio = "3.6.1"
async-timeout = "4.0.2"
click = "8.1.3"
colorama = "0.4.5"
Deprecated = "1.2.13"
fastapi = "0.78.0"
greenlet = "1.1.2"
h11 = "0.13.0"
idna = "3.3"
isort = "5.10.1"
Mako = "1.2.1"
MarkupSafe = "2.1.1"
packaging = "21.3"
psycopg2-binary = "2.9.3"
pydantic = "1.9.1"
pyparsing = "3.0.9"
redis = "4.3.4"
sniffio = "1.2.0"
SQLAlchemy = "1.4.39"
sqlalchemy2-stubs = "0.0.2a24"
sqlmodel = "0.0.6"
starlette = "0.19.1"
typing-extensions = "4.3.0"
uvicorn = "0.18.2"
wrapt = "1.14.1"
passlib = "^1.7.4"
alembic = "^1.8.1"
PyJWT = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
